Draft: Different expected output for invoke-test.ll on Clang 13
TODO RGS: Consider carefully if this is the approach we actually want to use.
Ultimately, the difference comes down to the fact that when `crux-llvm` calls
`llvm-link invoke-ll.bc -o invoke-ll.bc` in its linking step, `llvm-link-12`
will rearrange the order of unnamed globals in the file, whereas `llvm-link-13`
will not. I have no idea why this difference exists, nor how to search for the
commit that introduced this difference in behavior.

[ci skip]